docs: note with more compiling details
I was stumped for the expected path to compile these clojure files - my
guess is that devs using this are familiar with arcadia-unity and the
CLR itself, which perhaps implies what build directory is being referred
to, and how clojure could be arbitrarily run in a CLR context.

This documents a possible way to make this work for users trying to
export a godot game without learning all that trivia.